[MATHS] Fix a lot of things

+++ src/include/maths.h
@@ -1,4 +1,4 @@
 #pragma once
 
-int evaluate_maths(char *str);
+long long int evaluate_maths(char *str);
 char *get_maths(char *str);
